refactor(module:mention): refactor
